Profile
Added profile pictures and gifs
Added riends and friends request, added a page to track our friends when they change their username
Added blocked people
Added also the typing from Typescript in the back
Added an history of game on each user's profile

Few other unrelated fixes :
added tailwind.config.js
added an event for navigation
changed few unrelated lines
fixed chat persistence on main(force)
etc.

Also, there is chat's branch code. Thus the code has websockets and chat by the back. Those are made by sjean

+++ services/client/src/ts/html/events.ts
@@ -1,8 +1,9 @@
 
 
-import { goToURL, keyExist, resetReconnectTimer } from "./utils.js";
-import { htmlSnippets, setLanguage, selectLanguage, findLanguage } from "./templates.js";
-import { main } from "./app.js";
+import { encodeURIUsername, goToURL, keyExist, resetReconnectTimer } from "../utils.js";
+import { htmlSnippets, findLanguage, selectLanguage, assetsPath, setLanguage } from "./templates.js";
+import { main } from "../app.js";
+import { sendChatMessage, sendStatusMessage } from "../chat.js";
 
 /**
  * set all the events that the page need to work properly
@@ -11,10 +12,14 @@ export function setEvents(): void {
   let elem: any;
   const events = {
     "chat-input": setEnterEventChat,
-    "user-search": setEnterEventUsername,
+    "user-search": setMultipleEventUsername,
+    "pfp-form": setSubmitEventPfp,
     "login-form": setSubmitEventLogin,
     "register-form": setSubmitEventRegister,
+    "blocking request" : setClickEventBlockRequest,
+    "friend request" : setClickEventFriendRequest,
     "profile-form": setSubmitEventProfile,
+    "pfp-input": setChangeEventPfpInput,
     "go-to-profile": setClickEventProfile,
     "change-password-form": setSubmitEventPassword,
     "delete-account-form": setSubmitEventDelete,
@@ -56,10 +61,45 @@ function displayErrorOrAlert(form: HTMLFormElement, message: string): void {
   }
 }
 
+function needConnection(form: HTMLFormElement): boolean{
+  if (localStorage.getItem("token") === null) {
+    displayErrorOrAlert(form, "You are not connected");
+    return false;
+  }
+  return true;
+}
+
 /**
  * used by the login form, the default event can't be used in SPA (redirect)
  * @param form the said form element
  */
+function setSubmitEventPfp(form: HTMLFormElement): void {
+  form.addEventListener('submit', async (event: SubmitEvent) => {
+    event.preventDefault();
+    const formData = new FormData(form);
+    try {
+      const response = await fetch('/pfp', {
+        headers: {
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
+        method: 'PUT',
+        body: formData,
+      });
+
+      resetReconnectTimer(response.headers.get('x-authenticated'));
+      const result: {success?: boolean, message?: string} = await response.json();
+      if (!result.success) {
+        displayErrorOrAlert(form, keyExist(result, "message") ? selectLanguage(result.message) : 
+        `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
+        return ;
+      }
+      main(true);
+    } catch (error) {
+      displayErrorOrAlert(form, String(error));
+    }
+  });
+}
+
 function setSubmitEventLogin(form: HTMLFormElement): void {
   form.addEventListener('submit', async (event: SubmitEvent) => {
     event.preventDefault();
@@ -69,7 +109,10 @@ function setSubmitEventLogin(form: HTMLFormElement): void {
       const password = formData.get("password") as string;
       const response = await fetch('/login', {
         method: 'POST',
-        headers: {"Content-Type": "application/x-www-form-urlencoded"},
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
         body: new URLSearchParams({
           username: username, 
           password: password,
@@ -83,13 +126,13 @@ function setSubmitEventLogin(form: HTMLFormElement): void {
         `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
         return ;
       }
+      sendStatusMessage();
       main();
     } catch (error) {
       displayErrorOrAlert(form, String(error));
     }
   });
 }
-
 
 /**
  * used by the Password change form, the default event can't be used in SPA (redirect)
@@ -108,7 +151,10 @@ function setSubmitEventRegister(form: HTMLFormElement): void {
       }
       const response = await fetch('/register', {
         method: 'POST',
-        headers: {"Content-Type": "application/x-www-form-urlencoded"},
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
         body: new URLSearchParams({
           username: username,
           password: password,
@@ -122,6 +168,7 @@ function setSubmitEventRegister(form: HTMLFormElement): void {
         `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
         return ;
       }
+      sendStatusMessage();
       main();
     } catch (error) {
       displayErrorOrAlert(form, String(error));
@@ -136,13 +183,19 @@ function setSubmitEventRegister(form: HTMLFormElement): void {
 function setSubmitEventProfile(form: HTMLFormElement): void {
   form.addEventListener('submit', async (event: SubmitEvent) => {
     event.preventDefault();
+    if (needConnection(form) === false)
+      return;
     const formData = new FormData(form);
     try {
+      const username = formData.get('username') as string;
       const response = await fetch('/update', {
         method: 'PUT',
-        headers: {"Content-Type": "application/x-www-form-urlencoded"},
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
         body: new URLSearchParams({
-          username: formData.get('username') as string,
+          username: username,
           biography: formData.get('biography') as string,
         }),
       });
@@ -154,7 +207,7 @@ function setSubmitEventProfile(form: HTMLFormElement): void {
         `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
         return ;
       }
-      goToURL(`profile/${formData.get('username') as string}`);
+      goToURL(`profile/${encodeURIUsername(username)}`);
     } catch (error) {
       displayErrorOrAlert(form, String(error));
     }
@@ -169,6 +222,8 @@ function setSubmitEventProfile(form: HTMLFormElement): void {
 function setSubmitEventPassword(form: HTMLFormElement): void {
   form.addEventListener('submit', async (event: SubmitEvent) => {
     event.preventDefault();
+    if (needConnection(form) === false)
+      return;
     const formData = new FormData(form);
         try {
       const password = formData.get("password") as string;
@@ -179,7 +234,10 @@ function setSubmitEventPassword(form: HTMLFormElement): void {
 
       const response = await fetch('/password', {
         method: 'PUT',
-        headers: {"Content-Type": "application/x-www-form-urlencoded"},
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
         body: new URLSearchParams({
           password: password, 
         }),
@@ -192,9 +250,9 @@ function setSubmitEventPassword(form: HTMLFormElement): void {
         `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
         return ;
       }
-      const elem = document.getElementById("username");
+      const elem = document.getElementById("username-p1");
       if (elem && elem.hasAttribute("value"))
-        goToURL(`profile/${elem.getAttribute("value")}`);
+        goToURL(`profile/${encodeURIUsername(elem.getAttribute("value"))}`);
       else
         goToURL( );
     } catch (error) {
@@ -206,11 +264,16 @@ function setSubmitEventPassword(form: HTMLFormElement): void {
 function setSubmitEventDelete(form: HTMLFormElement): void {
   form.addEventListener('submit', async (event: SubmitEvent) => {
     event.preventDefault();
+    if (needConnection(form) === false)
+      return;
     const formData = new FormData(form);
     try {
       const response = await fetch('/delete', {
         method: 'DELETE',
-        headers: {"Content-Type": "application/x-www-form-urlencoded"},
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
         body: new URLSearchParams({
           username: formData.get("username") as string, 
         }),
@@ -223,6 +286,7 @@ function setSubmitEventDelete(form: HTMLFormElement): void {
         `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
         return ;
       }
+      sendStatusMessage();
       main();
     } catch (error) {
       displayErrorOrAlert(form, String(error));
@@ -232,16 +296,70 @@ function setSubmitEventDelete(form: HTMLFormElement): void {
 
 /**
  * used by the button that goes to the public profile from /profile
- * @param textElem the HTMLElement that's gonna get the event
+ * @param text the HTMLElement that's gonna get the event
  */
 function setClickEventProfile(text: HTMLElement): void {
-  const usernameElem = document.getElementById("username");
+  const usernameElem = document.getElementById("username-p1");
   if (!usernameElem || !usernameElem.hasAttribute("value"))
     return ;
   const username: string = usernameElem.getAttribute("value");
   text.addEventListener("click", (event: PointerEvent) => 
-    goToURL(`profile/${username}`)
+    goToURL(`profile/${encodeURIUsername(username)}`)
   );
+}
+
+function setClickEventBlockRequest(text: HTMLElement): void {
+  if (text.innerText === "NOT CONNECTED" || text.innerText === "IT IS YOU") {
+    text.innerText = findLanguage(text.innerText.toLowerCase());
+    return ;
+  }
+  let username = location.pathname.substring(location.pathname.lastIndexOf('/') + 1);
+  text.addEventListener("click", async (event: PointerEvent) => {
+    try {
+      const response = await fetch(`/block?user=${username}`,{
+        method: 'post',
+        headers: {
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
+    });
+      resetReconnectTimer(response.headers.get('x-authenticated'));
+      const result: {success?: boolean, message?: string} = await response.json();
+      if (!result.success) {
+        sendMessage(keyExist(result, "message") ? selectLanguage(result.message) : 
+        `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
+      }
+    } catch (error) {
+      sendMessage(String(error));
+    }
+    main(true);
+  });
+}
+
+function setClickEventFriendRequest(text: HTMLElement): void {
+  if (text.innerText === "NOT CONNECTED" || text.innerText === "IT IS YOU" || text.innerText === "THEY ARE BLOCKED") {
+    text.innerText = findLanguage(text.innerText.toLowerCase());
+    return ;
+  }
+  let username = location.pathname.substring(location.pathname.lastIndexOf('/') + 1);
+  text.addEventListener("click", async (event: PointerEvent) => {
+    try {
+      const response = await fetch(`/friend?user=${username}`, {
+        method: 'post',
+        headers: {
+          "Authorization": `Bearer ${localStorage.getItem("token")}`
+        },
+      });
+      resetReconnectTimer(response.headers.get('x-authenticated'));
+      const result: {success?: boolean, message?: string} = await response.json();
+      if (!result.success) {
+        sendMessage(keyExist(result, "message") ? selectLanguage(result.message) : 
+        `${findLanguage("server answered")} ${response.status} ${response.statusText}`);
+      }
+    } catch (error) {
+      sendMessage(String(error));
+    }
+    main(true);
+  });
 }
 
 /**
@@ -252,14 +370,10 @@ function setEnterEventChat(textarea: HTMLTextAreaElement): void {
   textarea.addEventListener("keydown", (event: KeyboardEvent) => {
     if (event.key === "Enter" && !event.shiftKey) {
       event.preventDefault();
-      if (textarea && textarea.value) {
-        sendMessage(textarea.value);
-        textarea.value = "";
-      }
+      sendChatMessage();
     }
   });
 }
-
 
 /**
  * to add a new message to the chat.
@@ -295,6 +409,11 @@ export function sendMessage(...args: any[]): boolean {
   if (scroll) {
     chat.scrollTop = chat.scrollHeight;
   }
+  if (chat.childElementCount > 1000) {
+    // we have to remove by an even (% 2) amount due to CSS even and odd colors 
+    chat.firstElementChild?.remove();
+    chat.firstElementChild?.remove();
+  }
   return scroll;
 }
 self["sendMessage"] = sendMessage;
@@ -304,14 +423,61 @@ self["sendMessage"] = sendMessage;
  * Preferably NOT DEFINITIVE (annoying to use)
  * @param textarea the said username input
  */
-function setEnterEventUsername(textarea: HTMLTextAreaElement): void {
+function setMultipleEventUsername(textarea: HTMLInputElement): void {
+  function clearSibling() {
+    const div = textarea.nextElementSibling as HTMLDivElement;
+    if (div) div.innerHTML = "";
+    first = null;
+    return div;
+  }
+  let first: string | null = null;
   textarea.toggleAttribute("data-custom", true);
   textarea.addEventListener("keydown", (event: KeyboardEvent) => {
     if (event.key === 'Enter' && textarea.value) {
       event.preventDefault();
-      goToURL(`profile/${textarea.value}`, true);
+      goToURL(`profile/${encodeURIUsername(textarea.value)}`, true);
       textarea.value = "";
+      clearSibling();
     }
+    if (event.key === 'Tab' && first !== null) {
+      event.preventDefault();
+      textarea.value = first;
+      clearSibling();
+    }
+  });
+  let searchController: AbortController | null = null;
+  textarea.addEventListener("input", async (e) => {
+    textarea.value = Array.from(textarea.value.matchAll(/[0-9a-zA-Z_:]/g), (m) => m[0]).join("");
+    try {
+      if (searchController) searchController.abort(); // cancel previous request
+      searchController = new AbortController();
+      const res = await fetch(`/misc/users?start=${textarea.value}`, {
+        signal: searchController.signal
+      });
+      const json = await res.json();
+      const div = clearSibling();
+      if (!div) return;
+      if (json[0]) first = json[0].username;
+      const fragment = document.createDocumentFragment();
+      for (let user of json) {
+        const span = document.createElement("span");
+        span.className = "flex flex-row cursor-pointer my-auto py-2 h-fit w-full gap-5 border border-black bg-gray-400";
+        span.addEventListener("pointerdown", () => goToURL(`/profile/${encodeURIUsername(user.username)}`, true));
+        const img = document.createElement("img");
+        img.className = "size-5 rounded-full";
+        img.src = `${assetsPath}/pfp/${user.pfp}`;
+        const p = document.createElement("p");
+        p.textContent = user.username;
+        span.append(img, p);
+        fragment.appendChild(span);
+      }
+      div.innerHTML = "";
+      div.appendChild(fragment);
+    }
+    catch (err) {};
+  });
+  textarea.addEventListener("blur", async () => {
+    clearSibling();
   });
 }
 
@@ -323,6 +489,26 @@ function setChangeEventLanguageSelector(select: HTMLSelectElement): void {
     localStorage.setItem("language", select.value);
     setLanguage();
     main(true, false);
+  });
+}
+
+function setChangeEventPfpInput(input: HTMLInputElement) {
+  input.addEventListener("change", (e) => {
+    const file = input.files[0];
+    if (!file) return;
+    
+    // Check MIME type
+    if (!file.type.startsWith('image/')) {
+      input.value = '';
+      document.getElementById("pfp-preview-div").toggleAttribute('hidden', true);
+      const form = document.getElementById("pfp-form") as HTMLFormElement;
+      if (form) displayErrorOrAlert(form, findLanguage("need image"));
+      return;
+    }
+    const img = window.URL.createObjectURL(input.files[0]);
+    const preview = document.getElementById("pfp-preview") as HTMLImageElement;
+    if (preview && img) preview.src = img;
+    document.getElementById("pfp-preview-div")?.removeAttribute('hidden')
   });
 }
 
@@ -353,16 +539,16 @@ export function setCtrlEventUsername(): void {
 
     if ((e.ctrlKey || e.metaKey) && ! e.shiftKey && e.key === 'p') {
       e.preventDefault();
-      const elem = document.getElementById("username");
+      const elem = document.getElementById("username-p1");
       if (elem && elem.hasAttribute("value")) {
-        goToURL(`profile/${elem.getAttribute("value")}`);
+        goToURL(`profile/${encodeURIUsername(elem.getAttribute("value"))}`);
         return ;
       }
       goToURL('profile');
     }
 
     if ((e.ctrlKey || e.metaKey) && e.key === 'e') {
-      if (self["isConnected"]) {
+      if (localStorage.getItem("token")) {
         e.preventDefault();
         fetch("/logout", {method: 'POST'}).then(
           res => {
@@ -372,6 +558,26 @@ export function setCtrlEventUsername(): void {
           }
         ).catch(err => alert('Caught: ' + err));
       }
+    }
+    
+    if ((e.ctrlKey || e.metaKey) && (e.key === 'ArrowRight' || e.key === 'ArrowLeft') && !isPressed && document.activeElement === document.body) {
+      const buttons = document.getElementById("inner-buttons");
+      if (!buttons)
+        return;
+      const checked = buttons.querySelector("[data-checked]");
+      let target: Element;
+      if (checked) {
+        target = e.key === 'ArrowLeft' ? checked.previousElementSibling : checked.nextElementSibling;
+        if (target === null)
+          target = e.key === 'ArrowLeft' ? buttons.lastElementChild : buttons.firstElementChild;
+      } else {
+        target = e.key === 'ArrowLeft' ? buttons.firstElementChild : buttons.lastElementChild;
+      }
+      if (target !== null && target instanceof HTMLElement){
+        target.click();
+        e.preventDefault();
+      }
+      return;
     }
   })
 
@@ -385,6 +591,7 @@ export function setCtrlEventUsername(): void {
       e.stopPropagation();
       // document active is set to body to not activate while using chat
       if (e.key === '?' && !isPressed && document.activeElement === document.body) {
+        help.innerHTML = "<p>" + findLanguage("pop up") + "</p>";
         isPressed = e.code;
         app.appendChild(help);
